merge seven seg files, documented functions
Merged seven_seg_tester.py and change_display.py files so that everything can be run off single file as required.
Added testing loop for Milestone 2 Part 2 which asks for message and time it should be displayed for.
Added function headers, file header and overall cleaned up code and comments.
Refresh rate added, hard-coded to 60hz for now. 
Option to 'blink' removed as it is not included in project specifications.